finance_chatbot_using_embeddings_/requirements.txt

streamlit
langchain
langchain-community
langchain-google-genai
langchain-groq
python-dotenv
chromadb
requests
